python/langchain-payment-agent/requirements.txt
# AgentGatePay LangChain Integration - Dependencies
# Python 3.12+

# Core AgentGatePay SDK (published to PyPI)
# Use >= to allow future patch versions
agentgatepay-sdk>=1.1.3

# LangChain framework
langchain>=0.1.0
langchain-community>=0.1.0
langchain-openai>=0.0.5

# Blockchain interaction
web3>=6.11.0
eth-account>=0.10.0

# Environment variables
python-dotenv>=1.0.0

# HTTP requests
requests>=2.31.0

# Flask for seller API
flask>=3.0.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-mock>=3.12.0
